Scss changes and Date testing
The date is acting strangely as are all the props where they are replicating twice and one being undefined. I am not sure the error as I can't find an answer online as to why either.  More try statements have been put in as it's only on first call not the second or any updates.